Update patch set 1
Patch Set 1:

Jens, thanks for the review. I've just split the patch from it's child to make it independent from that change.

Here's the failure where ROUTER_GW_IP doesn't get set and thus devstack fails:

http://logs.openstack.org/33/188033/4/check/gate-heat-dsvm-functional-orig-mysql/13705bd/logs/devstacklog.txt.gz#_2015-09-25_18_57_51_455


So it fails because of hash randomization in Python 3 making the output inconsistent: http://paste.openstack.org/show/475747/

The current awk expression failed for this. So I've updated the expression to account for this inconsistency.

Patch-set: 1
Label: Verified=0