Perform service disablement in a lazy fashion
With the advent of plugins and their settings files it has become
possible to disable_service in local.conf only to have the service
re-enabled in a plugin settings file. This happens because of
processing order.

To get around this the disable_service function now aggregates
requests into a DISABLED_SERVICES variable which is then processed
by perform_disable_service just after the "source" phase of
stacking. This allows local.conf to win the race which is likely to
more accurately reflect people's expectations.

Change-Id: I0f9403f44ed2fe693a46cd02486bd94043ce6b1a
Closes-Bug: #1504304