Specify the default v3.25 to RYU_BRANCH in case of the ryu plugin
Currently RYU_BRANCH variable is definded in stackrc by default master
and it is used by both ofagent and ryu plugin.
But in the latest version(3.26 and later) of ryu library, some rest
api modules to be used by ryu plugin were removed.
So we use refs/tags/v3.25 as default branch in the case of ryu plugin.

Change-Id: Ie244f610260560d8efb2fb54ef3ecbb7e11968b7
Closes-Bug: 1505561