Add script to create a subunit streams
This commit adds a generate_subunit.py script which is used to
generate a subunit stream with a single result for a period of time.
It takes it 2 mandatory args and 2 optional to specify the start time,
stop time, and optionally the status (it defaults to sucecss) and an id
for test (if one isn't provided 'devstack' is used) The resulting stream
is written to STDOUT.

The intent here is to leverage this to inject 'test results' into the
subunit2sql db to reflect failures that occur before tempest (or any
other test suite) is run. This is necessary for the openstack-health
dashboard. (otherwise it makes our failure rate look much better than
it is) This is only needed until we get a zuul mysql reporter in place
which can give us the higher level run information.

Change-Id: I6bce2c76bbf85ea558a12ba673455332a862a4f9