Add subunit output to the timing infrastructure
This commit adds writing a subunit file to the timing infrastructure
in devstack. This means in addition to printing to stdout at the end
of a run, there will be a subunit stream available at the end of the
run which will contain all the individual timing data which can be
used for trend analysis with a subunit2sql db or visualized with
stackviz.

Change-Id: I4591cad7f353e070d21218843bad2f762870df21