Update patch set 2
Patch Set 2:

Hi Joseph,

The 'run_validation' flag is 'False' by default and because of this there are some tests that not running.
In the previous patch set I set this flag to True and only gates with nova network failed.
This why I'm trying to enable this flag only for neutron.

Patch-set: 2