Update patch set 1
Patch Set 1: Code-Review-1

This is not the way to go. We should not ask for each service to kill itself in the cleanup phase. We have generic mechanisms (screen_stop_service and/or stop_process) to do that. If there's a bug in these, we should clearly explain what's going on, including a reproducer. Other we will end up with a lot of code duplication

Patch-set: 1
Reviewer: Gerrit User 7350 <7350@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1