Update patch set 1
Patch Set 1:

On reviewing the existing comment about operating system and version, step 1 of http://docs.openstack.org/developer/devstack/#quick-start , I see that it does not now match what I was remembering.  Regarding the current statement: it is silent about later versions of Ubuntu and the redhatverse.  I am suggesting it would be helpful to specifically *dis*courage people from trying OpSys.versions that we know will *NOT* work.  Some of us are doing work that goes better with more recent kernels, so we are motivated to try later opsys versions if there is a possibility that it might work.

It may be true that statements of what *does* work can rot, (a) I am not suggesting we add any and (b) there already are some (thankfully!).

Patch-set: 1