Introduce IMAGE_PROPERTIES for custom image properties
PPC64LE architecture requires different image properties
for third party CI testing. The custom configuration
is usually set on devstack-gate hooks in order
to customize the deployment and tests.

Rather than adding hardcoded image properties for a new
architecture, the new DevStack property in localrc allows
the image properties to be overwritten not only for
production CI but also for eventual testing and custom
environments.

Closes-Bug: #1509038

Change-Id: Ia058e33ecef8ffe02ca7fc7119537b2ca06718eb