Update patch set 2
Patch Set 2: Workflow-1

Thanks, I will separate it out into two different things. When a real bare metal is used with tempest, we need to change how much time it takes to provision the system. It might range from 15-20 because it includes reboot of the servers.  Also, since this is dependent on the server itself, we needed it to be configurable.

Patch-set: 2
Label: Workflow=-1