Update patch set 1
Patch Set 1: Code-Review-1

I'm not sure Ian. What I've seen is a lot of related bugs saying that neutron doesn't starts when this option was activated. When using wget to check the neutron connection, the command doesn't returned anything but by using curl I saw that the problem was with the certificate validation.

Patch-set: 1
Label: Code-Review=-1