Update patch set 2
Patch Set 2:

Ian, so I agree that we don't need the validation in here, the config option in nova uses the choices kwarg which will raise a ValueError if you specify an invalid value:

https://github.com/openstack/nova/blob/master/nova/virt/libvirt/storage/lvm.py#L35

However, the default in nova is 'zero' and we need to change that to 'none' for the CI job runs to avoid any overhead messing up the test time.

Maybe it'd be simpler if we just assume devstack is not used for production and set it to 'none' explicitly in devstack if NOVA_BACKEND == 'lvm'?  What do you think about that?

Patch-set: 2