Ensure single ceph osd is weighted correctly
This patch ensures the OSD has a weight greater than 0 after deployment.
Ceph will not place data on an unweighted OSD and since there is
typically only one OSD for a devstack deployment, it makes the Ceph
cluster unusable.

As for why this behaviour changed recently, I don't have an answer.  But
generally ensuring a reasonable OSD weight is something we should be
doing anyway.

Closes-Bug: #1475333
Change-Id: Ied44ed81d890d25c43e563e844479924ac5f4a04