Update patch set 3
Patch Set 3: Code-Review-1

assumptions are always wrong for someone so we shouldn't make them.  having an override is ok, but by default we should leave it alone.

devstack-gate is an appropriate place for policy like this, however.  in this case, we already have sections in devstack-gate where we configure different things for different libvirt drivers [1].  it feels like it would be quite reasonable to specify defaults for lxc jobs in there.  in fact, it's probably reasonable to set the override generically in there for all gate jobs?

another option is passing it through for individual jobs in the job definition with DEVSTACK_LOCAL_CONFIG

[1] http://git.openstack.org/cgit/openstack-infra/devstack-gate/tree/devstack-vm-gate.sh#n318

Patch-set: 3
Label: Code-Review=-1