Update patch set 3
Patch Set 3:

Yeah so we can use DEVSTACK_LOCAL_CONFIG, that's easy, but we still need the var in devstack to set that will then set the option in nova.conf, which is what was initially in this change, minus the value validation - are you saying it's fine then to go back to that (minus the validation part)? I feel like we're going in circles on this one.

Patch-set: 3
Label: Verified=0