Update patch set 5
Patch Set 5:

Thanks; the code looks like I would expect.

IMO, the idea that we have to specify this for basically every user of oslo.service seems to point us in the direction that this is not a great idea.

When the default behaviour seems to be "send it SIGTERM, and it probably won't actually stop, because the children are busy for an unbounded amount of time ... oh, you really want it to shutdown ... you should have set this variable here"

when everyone is trying to ansible & puppet everything, stuff like services that don't stop are a real pita

Patch-set: 5
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0