Update patch set 5
Patch Set 5:

Hi Ian Wienand,
My point of view is that we don't need to specify this for basically every user of oslo.service.
I believe that we don't need it if graceful termination of service works fine and client closes all connection gracefully.
So if services and clients work well we don't need it.

If service is terminated because of graceful shutdown timeout we will have message https://github.com/openstack/oslo.service/blob/master/oslo_service/service.py#L344 in log file and it could be a reason to report a bug(that graceful shutdown doesn't work)
This message could be changed to warning...or even error message.

This review allows to close bug #1446583 but at the same time graceful shutdown problems will be visible.

What is your point of view?

Patch-set: 5