Update patch set 7
Patch Set 7: Code-Review-1

We have a non-voting job with ssh validation on running on check, but because we don't collect test results from check we have not been able to asses its stability yet.
There's a few subunit2sql performance issues we have to solve before we can start to collect data from the check queue. 
I feel a bit uncomfortable changing this without having an idea of the stability of the job, or no-one having really looked into the failures of this job. 
It's true however that there's a lot of noise in the check queue, so switching this one might be the right approach - and if so we should do that in the early stages of Mitaka - but be ready to revert if this affects gate stability.
I would rather see the ssh job made voting than this changed for every job as it will make it easier to identify and isolate ssh specific issues. If there is no issue, we can do this as a next step, and drop the ssh job completely.

Patch-set: 7
Reviewer: Gerrit User 1921 <1921@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1