Update patch set 10
Patch Set 10:

Patch Set 10 spawns neutron-api, and neutron-agent correctly. However, it is not flexible at this point. Hard coded choices on the ML2 plugin and Open vSwitch agent, but no calls into lib/neutron_plugins/ to configure things. Still a lot of work to do, a big thing is making the new neutron library compact, with just enough functionality to launch a guest instance on a flat network, and then enough hooks for Neutron DevStack plugins so that more complex things are possible.

Patch-set: 10
Label: Verified=0