Update patch set 4
Patch Set 4: Code-Review+2 Workflow+1

It isn't deployers that'll have a problem with this on by default, is us poor folk primarily running DevStack in 4Gb vms on laptops.  This is the death-blow for that with a default config.

I have a couple of follow-up tweaks to make but nothing critical to hold this up...

Patch-set: 4
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2
Label: Workflow=+1