Update patch set 1
Patch Set 1:

so, I realize I started this trend, but I'm actually a little conflicted at the moment. I sort of like the idea of actually moving the general stuff for things like dev packages back to the places that actually require it. Because if we ever want to get rid of lxml, we're going to need to make it clear where it's used, and where it can be removed.

I think one of the missed places was actually n-cpu for lxml.

Patch-set: 1
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0