Update patch set 1
Patch Set 1: Code-Review+2

after a bit of research, it seems to be upstart that does this only, so i guess we don't need the rpm-distro branching, but i guess it doesn't hurt either.

also seems the simpler idea of just making it "restart" doesn't work with upstart either if it's not already running

presumably the package install is starting the service, which is where this is coming from?  not sure why we didn't notice till now?

Patch-set: 1
Label: Code-Review=+2