Update patch set 2
Patch Set 2: Code-Review-1

This takes an assumption that zookeeper would only be used in devstack as a part of DLM, which is not a true.  Cue and Monasca use zookeeper for non DLM purposes.  For any project that hopes to use task flow job boards (cue currently, possibly others) or any other services that depend on zookeeper (Monasca uses Kafka which depends on Zookeeper) DLM is not a requirement for zookeeper, but zookeeper is a requirement for other reasons.  Zookeeper should be kept a separate thing that can be enabled irrespective of DLM.

Patch-set: 2
Reviewer: Gerrit User 10584 <10584@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1