Update patch set 2
Patch Set 2: Code-Review+1

For me this it is good, but I still prefer the more generic solution of patch-set 1. If starting a service twice fails on Ubuntu, than the start_serice function is broken and it has to be fixed.

Who whants to workaround this problem it is possible disabling zookeeper service as I wrote on the bug report in launchpad.

I sudgest to merge this patchset-2 (less risky) and then start investigating on the impact of applying patchset-1 on other distros for a global fix of start_service function.

I am going to test patchset-2 as I did for patchset-1.
Thanks for it.

Patch-set: 2
Reviewer: Gerrit User 17377 <17377@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1