Update patch set 5
Patch Set 5:

My worry here, and that I see discussed in some places but not really resolved, is that (dlm == zookeeper) seems correct -- devstack can be opinionated and say "if you're doing dlm work and you call use_dlm you're going to get zookeeper".  That said, obviously the whole idea of abstracting implementation behind a "dlm" alias is that you retain the possibility of switching it for something else.  The idea is that people are using tooz and shouldn't even really care what's underneath.  

But the reverse, (zookeeper == dlm) is not true.  There are examples of zookeeper being used for service discovery and coordination. So if these services call "use_dlm" just to get zookeeper (everyone has to coexist) we've just broken the whole point of the abstraction.

My suggestion would be to retain the zookeper service -- it seems like multiple projects want it so centralising install/setup of it is good -- and make the "dlm" service an alias to zookeeper.

Patch-set: 5
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0