Update patch set 5
Patch Set 5:

This is all just a terminology fight right? There is literally no code difference between this than the name.

If someone else has a better generic name than dlm for class of things here, lets use that. But we don't even call rabbit "messaging" in openstack, we call it rpc_backend. Even though it's used for notifications by as many things as rpc. 

I really think that zookeeper in devstack needs to always be behind an abstraction, or never be behind and abstraction, and we give up on the idea that tooz is our interface. This middle ground is just a mess for deployers and developers.

We blessed this as a thing in OpenStack you could depend on based on a specific abstraction for it.

Patch-set: 5
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0