Update patch set 4
Patch Set 4: Code-Review-1

honestly, I feel like this is going to get folks into more trouble. They rebase a change and forget that they have to update this hash and get stuck on the old one.

Building a local branch for a change is really just a better model, as it supports rebasing (which happens for a lot of reasons).

Patch-set: 4