Update patch set 4
Patch Set 4: Code-Review+1

this would speed up the intel networking ci as we would only have to clone the delta instead of reclone the entire repo for our dependencies. 

i agree that if you are developing the project that you are cloning then a branch would be better but in this case we are using it to manage a dependency and pin it to a know working commit. with this change we can easily move to another pinned commit of our dependency simply by updating the default in our devstack plugins settings file.

Patch-set: 4
Reviewer: Gerrit User 11604 <11604@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1