Update patch set 4
Patch Set 4: Code-Review-2

But you can already do that by managing local mirrors of whatever that is, there is no reason that a git url needs to be upstream.

I don't want to do this because the top place where people use this functionality is in openstack projects, testing a commit. And the failure mode when the update their branch and don't remember they pinned incorrectly is really really weird, and something we're going to have to debug for people regularly.

For CI speedups just cache the base tree locally, that's what upstream CI does with devstack-gate.

Patch-set: 4
Label: Code-Review=-2