Update patch set 4
Patch Set 4:

i don't want to push this too much so if you think it will result in more debugging then thats fine, i understand you don't want to introduce new failure modes and that is a valid point. the current failure mode of it works if the repo is not there was equally annoying to debug.

honestly the use of a commit id is identical to using a tag which is supported. if you forget to update the tag you will have the same weird behavior when you rebase. after all a tag is just a label for a commit.

we can continue to use our current logic of delete repo and let devstack reclone or implement it in our devstack plugin locally. we just taught this was an optimisation that was nice to share. we have local mirror of all openstack projects but not of ovs or dpdk, we are in the process of setting those up too though mirrors will not help with pinning external dependencies.

Patch-set: 4