Neutron: Fix provider network example in doc
The provider network example in the Neutron guide sets FIXED_RANGE with out
setting NETWORK_GATEWAY.  If NETWORK_GATEWAY is not specified, stack.sh will
use the default value of 10.0.0.1 which will result in the following error:

   "Invalid input for operation: Gateway is not valid on subnet."

IP_VERSION is set to "4" to get around a devstack IPv6 error with proivder
networks.  This secondary issue will be fixed in a separate patch.

Change-Id: Ifc709d68f9d4ac787214d3a60a86a52b7c2437d3