Update patch set 14
Patch Set 14:

Hi Hirofumi,

I see your comment now after Ian's. Good point, thanks.

It would seem then that LinuxBridge and OpenvSwitch plugins are treated as ML2 agents. I do not see an mention of "Sriov" in the DevStack.

The way the implementation is at the moment is that the sample version of the config files will be generated as mentioned in the comment above. The changes in this patch will however handle agents that exist as ML2 agents and if plugins have config files without ".sample" postfix. Maybe some of the code in this area is redundant?

I would appreciate any further feedback on this from you or anyone else.

Thanks,
Martin

Patch-set: 14