Update patch set 14
Patch Set 14:

Ian, the fact that devstack configures OVS and LB agents using ml2_conf.ini is a bug in devstack. Instead, devstack should use *_agent.ini files for agents, and use ml2_conf.ini for neutron-server only.

That said, this bug is out of scope for the patch.

Patch-set: 14