Update patch set 4
Patch Set 4:

> Why are we re-naming variables like PROVIDER_SUBNET_NAME_V6 to IPV6_PROVIDER_SUBNET_NAME? Can't we just introduce defaults?

PROVIDER_SUBNET_NAME_V6 is specified, but it was never assigned a value.  All of the rest of the IPv6 properties used IPV6_  as a prefix, rather than _V6 as a suffix.  I was trying to be consistent with the rest of the IPv6 properties.

V6_NETWORK_GATEWAY is specified, but it was never assigned a value.  Looks like IPV6_PRIVATE_NETWORK_GATEWAY is what was actually desired in that case.

Patch-set: 4
Label: Verified=0