Neutron: Create move_ip.sh tool
When using Neutron, L3 agent, and OpenVSwitch, the IP address is moved from
the physical interface, ethX, over to the OVS bridge, br-ex.  This is done
by ./stack.sh during the neutron L3 configuration.  This movement is
required in order to maintain connectiviy to this IP address.

When the host machine is rebooted OR the network service is restarted, the
IP address is associated with the physical interface, ethX, once again and
the IP address is no longer reachable.

The move_ip.sh tool provided in this patch can be used to move the IP address
from the physical interface back to OVS bridge.  The settings in local.conf
are used to validate that the neutron L3 service is enabled AND the values
to use for the physical interface and OVS bridge name.

Closes-Bug: #1521790
Change-Id: I0c80bb79c2bd1b5098f6dda0803798fed9d13463