Neutron: Add rejoin_neutron function
When using Neutron, L3 agent, and OpenVSwitch, the IP address is moved
from the physical interface, ethX, over to the OVS bridge, br-ex. This
is done by ./stack.sh during the neutron L3 configuration.  This
movement is required in order to maintain connectivity to this IP
address.

When the host machine is rebooted OR the network service is restarted,
the IP address is associated with the physical interface, ethX, once
again and the IP address is no longer reachable.

Update rejoin.stack.sh to move the IP address (if necessary) using
the same logic that invoked during stack.sh

Closes-Bug: #1521790
Change-Id: I0c80bb79c2bd1b5098f6dda0803798fed9d13463