Add distro support for KVM for IBM z Systems
"KVM for IBM z Systems" does not ship with a /etc/redhat-release
file. If this file does not exist, use /etc/base-release instead
which contains the relevant vendor and version info.

Also add "KVM for IBM z Systems" to the list of Fedora-like distros.

As the distribution does not have a dedicated kvm package,
prevent the installation of the kvm package during the libvirt
setup.

Change-Id: Ibb5c60797d6867264f9dea7fea85cdf1d7c72ded