Update patch set 5
Patch Set 5: Code-Review-1

No. Why would we have provider networking set - and then use IPV6_PRIVATE_NETWORK_GATEWAY for the gateway? The gateway for provider networking will be the LLA of a hardware switch. By default IPV6_PRIVATE_NETWORK_GATEWAY is set with a unique local address prefix that we generate randomly.

V6_NETWORK_GATEWAY was specifically intended to be defined by the user who will know the LLA of their hardware router, if they are using this feature. Perhaps it just requires more documentation on my part since it took a while for me to remember why we have it.

Patch-set: 5
Label: Code-Review=-1