Update patch set 5
Patch Set 5:

Ok, so V6_NETWORK_GATEWAY should be set to the LLA from the provider.  This is not documented in the provider network section of the neutron guide. 

Using the properties that are documented, the neutron subnet-create on line 550 fails because V6_NETWORK_GATEWAY is not specified. 

The neutron subnet-create that is done on line 550 specifies FIXED_RANGE_V6 therefore I changed the gateway to match that.

Seems like it should not be using FIXED_RANGE_V6 since that is used for the private network.

Patch-set: 5