Update patch set 9
Patch Set 9: Code-Review+1

federico correct me if im wrong but federico orignially discovered this bug while working on 
https://review.openstack.org/#/c/218210/

in the networking-odl plugin it was being called without set -e.
federico was testing the return value of install_package 
when he discovered that the behaviour changed after the timing info was added to apt_get.

Patch-set: 9