Update patch set 9
Patch Set 9:

Jordan, this bug is a regression, the change I propose is to restore original behaviour. The bug is explained on launchpad:

https://bugs.launchpad.net/devstack/+bug/1518545

'set -e' is not taken in consideration when called function is part of a logical expression.

For example before this change it would never enter inside of this condition:

if ! apt_get peter-pan; then
    echo 'Unable to install peter-pan'
fi

The same bug also affect install_package (only on ubuntu) and pip_install (all distributions).

As a general rule I think we should avoid assuming a library has different behaviour when is called in different contexts.

this script is a library and it should behave always the same without depending on which script sources it.

Patch-set: 9
Label: Verified=0