Update patch set 1
Patch Set 1:

Hi Dean,

this change is about running devstack on "KVM for IBM z Systems", which is a hypervisor distro available for s390x. So it's not really about Fedora.

More info is available here: http://www-03.ibm.com/systems/z/solutions/virtualization/kvm/

The distro does not contain a /etc/redhat-release file, but uses /etc/base-release for the same kind of information. We also don't have lsb_release right now so the extensions suggested in the patch are the only viable solution I found so far.

Patch-set: 1