Update patch set 1
Patch Set 1:

Running unstack.sh is a alternative solution, however it requires that the system be re-installed via stack.sh and reconfigured (networks, floating IPs, VMs, load balancers, etc... ).  This reconfiguration could be extensive.

The rejoin-stack.sh seems like it was designed to handle recovery after a reboot of the host system, however it lacks the logic to re-establish network connectivity to the physical interface for this specific case.

Patch-set: 1
Label: Verified=0