Update patch set 2
Patch Set 2:

Matthew,

Generally I disagree. Anyone running a traditional devstack installation won't want disk erasure enabled as implemented by Ironic -- it'll fall back to a shred which can take hours, and waste write-lifetime on your machine you're testing on.

The "real fix" is a clean step that runs in devstack and confirms the coordination around cleaning is working, even if the steps themselves are not hardware-specific.

Patch-set: 2