Update patch set 2
Patch Set 2:

We've been discussing an approach like what Jay describes for testing this in the gate.

However, I still think that we want this to default to run cleaning in a stock devstack install, once we have a way to verify the coordination in the gate.

Lots of folks try out devstack first and it should not default to leaving baremetal hosts intact after instance deletion, just because the alternative is slower.

Patch-set: 2
Reviewer: Gerrit User 2889 <2889@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0