Update patch set 10
Patch Set 10: Code-Review+1

I think this is a great cleanup and puts a good opinion into the code: you need a sane environment to run this thing. That's okay given the meaning and import of devstack these days (which is not the same as it ever was).

Patch-set: 10
Label: Code-Review=+1