Update patch set 1
Patch Set 1:

Hi Ian,

I tried to contact you on #openstack-dev to ask for more details on your proposed solution. Do you suggest to add a custom lsb_release script to the devstack repository that would be called if lsb_release is not available? That would serve as last resort in your _ensure_lsb_release() function if there is no known package that can be installed to get lab_release?

If my understanding is correct, I can look into providing this kind of script.

Patch-set: 1