Update patch set 1
Patch Set 1:

unstack.sh can not be run after the reboot if network connectivity was lost and you have no other means to access the host system.   The script in this patch could be added to the startup logic such that network connectivity is automatically restored after a reboot.

Patch-set: 1