Cap tox installation at !=2.3.0 to fix ValueError
Tox 2.3.0 released today and is causing a "ValueError: need more
than 1 value to unpack" when trying to parse tempest's tox.ini.
This is a temporary solution which blacklists Tox 2.3.0 so
stacking does not fail and should be reverted as soon as the fix
for Tox is released.

(Note that this issue only affects individuals and 3rd party CIs
running Devstack and that the gate does not see this problem
since Tox is installed via other means.)

Closes-Bug: #1525352
Change-Id: I46e14b15532b41ff2db2b61ff8ebcf545cb28db6