Update patch set 2
Patch Set 2: Code-Review-1

Do we need to have backwards compatibility with enable_service q-vpn? I would prefer to see us fix it at the gate (if it is still being used in a job configuration in project-config) and emit a warning or error that q-vpn is deprecated. 

The reason I say this is because I don't like having DevStack injecting more things that are defaults. What if the defaults change in the future? Wouldn't we have to update the defaults again in DevStack? Plus, VPNaaS was moved to a plugin - so let's go ahead and cut the umbilical cord?

Let me know if I'm my thinking is wrong, I do know this is causing us grief

Patch-set: 2