Update patch set 3
Patch Set 3:

I hate to be nit-picky, but pbr isn't meant to have a dep on testrepository - its not *missing*. There is a bug in old pbr, which kilo decided to stay capped on due to the effort required to fix /all of kilo/ to not accept newer pbr in its versions, and a separate bugfix in setuptools has exposed this old pbr bug - which isn't present in current pbr. The workaround here is ok, but we should press ahead and do that point release.

Patch-set: 3
Reviewer: Gerrit User 4190 <4190@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0