Update patch set 10
Patch Set 10:

Jordan, this bug fix is not because of a bad side effect. This correct the logic expression that make devstack always trying to install a package again once it was did before. On the case of failure current behaviour is right, but on the case on successful installation package is tried to install again wasting time. This is easy to demonstrate executing a simpler bash expression as explained in the bug description.

Patch-set: 10