neutron: enable MTU advertisement
This is needed if you want to run devstack using tunnelling for tenant
networks as well as multiple nodes. In this case, packets get additional
encapsulation header, and if packets are of physical device mtu size,
tunnelled data can be dropped.

MTU is advertised using DHCP and RA payloads. It works as long as
guests know how to handle the settings. (Cirros honors DHCP option as of
0.3.4 release).

Note that this could also be solved by setting the default value for the
option to True on neutron side. But we'll  also need to backport the
change to stable/liberty to make multinode grenade job passing proper
mtu value into instances created on the 'old' side of grenade. Since a
patch that would change the default value for the option would not
conform to stable maintenance requirements, this patch is needed, at
least in stable/liberty branch of devstack.

Change-Id: Idc18dcf479e0ac6d293c2cbee29a550601a1fa22
Related-Bug: #1527675