neutron: enable MTU advertisement
This is needed if you want to run devstack using tunnelling for tenant
networks as well as multiple nodes. In this case, packets get additional
encapsulation header, and if packets are of physical device mtu size,
tunnelled data can be dropped, especially when "Dont Fragment" (DF) flag
is set.

MTU is currently advertised using DHCP option only. It works as long as
guests know how to handle it. (Cirros honors the option as of 0.3.4).

Note this patch is Liberty only. For Mitaka, we plan to change the
default value for the option to True (see the dependency). We won't be
able to backport the neutron patch to stable release since it violates
stable maintenance rules.

Change-Id: I53c0a29438d0d6433efd4b379bc0391a69d26589
Depends-On: I5cbbc4660f8c4e15e59f8f5ce0419501bdd27348
Related-Bug: #1527675