Fix ordering of check in yum_Install, so PIPESTATUS[0] is checked first
Re-ordered check for how yum_install went, so PIPESTATUS[0] is the value
of the awk command. Failure in doing so renders the check to be after
the if statement, which gives it the value of 1.

This issue apperas to have been introduced in commit 2babf396

${PIPESTATUS[0]} != 0

...
Running transaction check
Running transaction test
Transaction test succeeded
Running transaction
  Installing : epel-release-7-5.noarch                                      1/1
  Verifying  : epel-release-7-5.noarch                                      1/1

Installed:
  epel-release.noarch 0:7-5

Complete!
+ result=0
+ '[' 0 '!=' 0 ']'   <== result
+ [[ 1 != 0 ]]       <== ${PIPESTATUS[0]}, used one statement too late
+ echo 1344 'yum install failure'
1344 yum install failure
+ result=1
+ return 1
+ die 286 'Error installing EPEL repo, cannot continue'
+ local exitcode=1

Reference-Change-Id: Ia33bb98f47017f6ec00eb3b7261ed42d72efc573
Change-Id: I801100134754454f902cfe9034021837b1a185dc
Closes-Bug: #1522590
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>