Update patch set 7
Patch Set 7:

It didn't seem to be a particular package that triggered this issue. My theory is something in the VM did this, but I'm not sure what exactly. I spent over a day on debugging this and I rebuilt the VM again and again and it would still happen. In my debugging I tried having Ansible install all the packages that devstack would initially install and so it went past that point but later when it went to install rabbitmq it hung. This patch fixed the issue.

Patch-set: 7