Switch tempest to use the public network
Tempest isn't guaranteed to have access to a private network. It
depends on the underlying network topology of the cloud and in many
cases the private network isn't externally accesible. Granted these
are just labels, but the implication of the name is we should be
using the public network, not the private.

Change-Id: I377479418b72464504863304a3a13f2721b4749d