Update patch set 1
Patch Set 1: Code-Review+1

On our multinode test setups using nova net we enable multi host because not using multi host was broken and no one was interested in fixing it. This means that each node in the test is its own router for the VMs. A consequence of this is that traffic from node A to a VM on node B over the private net isn't routable. You can get lucky and have traffic from node A hit VMs on node A over the private net but not across nodes.

To solve this general connectivity problem we put every hypervisor on the public net that floating IPs are attached to. This means that any hypervisor can talk to any VM over the floating IP over the public network.

Patch-set: 1
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1