Update patch set 1
Patch Set 1:

I think in your bug report, part of the arguments that are passed to radvd incude the string "/opt/stack/data/neutron/" - IIRC the -f param to pkill allows you to add more search criteria for killing a process.

As for root - sure, you'll still need to sudo. But do you need to SIGKILL it? What about just doing SIGTERM to let it clean up and exit?

Patch-set: 1