Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

i have a few in line comments, but i'm wondering if we can get most of the necessary changes in without depending on the ironic migration to keystoneauth? i think only the change in lib/nova_plugins/hypervisor-ironic is needed for ironic right?

Patch-set: 2
Label: Code-Review=-1